+++ src/trx/game/rooms/geometry.h
@@ -17,12 +17,12 @@ bool Room_IsAbyssHeight(int32_t height);
 HEIGHT_TYPE Room_GetHeightType(void);
 int16_t Room_GetTiltType(const SECTOR *sector, XYZ_32 pos);
 
-int16_t Room_GetHeight(const SECTOR *sector, XYZ_32 pos);
-int16_t Room_GetHeightEx(
+int32_t Room_GetHeight(const SECTOR *sector, XYZ_32 pos);
+int32_t Room_GetHeightEx(
     const SECTOR *sector, XYZ_32 pos, bool fix_tilts, int16_t ignore_item_num);
-int16_t Room_GetCeiling(const SECTOR *sector, XYZ_32 pos);
-int16_t Room_GetCeilingEx(const SECTOR *sector, XYZ_32 pos, bool fix_tilts);
-int16_t Room_GetFloorHeightForSector(
+int32_t Room_GetCeiling(const SECTOR *sector, XYZ_32 pos);
+int32_t Room_GetCeilingEx(const SECTOR *sector, XYZ_32 pos, bool fix_tilts);
+int32_t Room_GetFloorHeightForSector(
     const SECTOR *sector, int32_t x, int32_t z, bool fix_tilts);
 
 int32_t Room_GetWaterHeight(XYZ_32 pos, int16_t room_num);
